fix some bugs and add success message

+++ src/main/java/com/openwt/boats/controller/BoatControllerImpl.java
@@ -1,10 +1,16 @@
 package com.openwt.boats.controller;
 
+import com.openwt.boats.authentication.Authenticator;
 import com.openwt.boats.dao.BoatRepository;
+import com.openwt.boats.dao.UserRepository;
 import com.openwt.boats.dto.boat.BoatRegisterRequestDto;
 import com.openwt.boats.dto.boat.BoatRegisterResponseDto;
 import com.openwt.boats.dto.boat.BoatResponseDto;
 import com.openwt.boats.entity.Boat;
+import com.openwt.boats.entity.User;
+import com.openwt.boats.error.UserError;
+import com.openwt.boats.exception.NotAuthorizedException;
+import com.openwt.boats.session.SessionInfo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Component;
@@ -23,6 +29,12 @@ public class BoatControllerImpl implements BoatController{
     @Autowired
     BoatRepository repository;
 
+    @Autowired
+    private Authenticator authenticator;
+
+    @Autowired
+    private UserRepository userRepository;
+
     public BoatRegisterResponseDto registerBoat(@RequestHeader UUID authToken,@RequestBody BoatRegisterRequestDto boatRegisterRequestDto) throws Exception {
 
         Boat boatEntity= new Boat(boatRegisterRequestDto.getName(),
@@ -35,7 +47,14 @@ public class BoatControllerImpl implements BoatController{
     }
 
     @Override
-    public BoatRegisterResponseDto updateBoat(@RequestBody BoatRegisterRequestDto boatRegisterRequestDto) throws Exception {
+    public BoatRegisterResponseDto updateBoat(@RequestHeader UUID authToken,@RequestBody BoatRegisterRequestDto boatRegisterRequestDto) throws Exception {
+
+        //Get Active Session
+        SessionInfo sessionInfo = authenticator.checkUpdateSession(authToken);
+        User user = userRepository.findUsersById(boatRegisterRequestDto.getUserId());
+        //Validate Authorization
+        if (!user.getId().equals(sessionInfo.getUserId()))
+            throw new NotAuthorizedException(UserError.UNAUTHORIZED);
 
         Boat boatEntity;
         if (boatRegisterRequestDto.getId()==null)
@@ -89,7 +108,15 @@ public class BoatControllerImpl implements BoatController{
     }
 
     @Override
-    public BoatRegisterResponseDto deleteBoat(@PathVariable Long boatId) throws Exception {
+    public BoatRegisterResponseDto deleteBoat(@RequestHeader UUID authToken,@PathVariable Long boatId,@PathVariable Long userId) throws Exception {
+
+        //Get Active Session
+        SessionInfo sessionInfo = authenticator.checkUpdateSession(authToken);
+        User user = userRepository.findUsersById(userId);
+        //Validate Authorization
+        if (!user.getId().equals(sessionInfo.getUserId()))
+            throw new NotAuthorizedException(UserError.UNAUTHORIZED);
+
         repository.delete(boatId);
         BoatRegisterResponseDto response = new BoatRegisterResponseDto(HttpStatus.OK,"Boat is deleted succesfully");
         return  response;
